dp git initialization

Initializes the git repo according to the requirements of dp

dp_git_init(
  project_path,
  project_name,
  branch_name,
  github_repo_url,
  board_params_set_dried,
  creds_set_dried,
  git_ignore
)

Arguments

project_path
Path to the project folder

project_name
The name of the project. This is typically the name of the folder where the project is set

branch_name
An abbreviation to capture the specific reason for which data was processed. Example m3cut (as in month 3 data cut)

github_repo_url
the https url for the github repo

board_params_set_dried
Character representation of the function for setting board_params. Use fn_dry() in combination with dpi::board_params_set_s3 or dpi::board_params_set_labkey. See example.

creds_set_dried
Character representation of the function for setting creds. Use fn_dry() in combination with dpi::creds_set_aws or dpi::creds_set_labkey. See example

git_ignore
A character vector of the files and directories to be ignored by git.